Created a working part of Factory Method pattern

+++ SquareParticle.cpp
@@ -1,6 +1,6 @@
 #include "SquareParticle.h"
 
-SquareParticle::SquareParticle(float speed, sf::Vector2f direction, float size, sf::Texture& texture)
+SquareParticle::SquareParticle(int speed, sf::Vector2f direction, int size, sf::Texture& texture)
 	: Particle(speed, direction, new sf::RectangleShape{ sf::Vector2f(size, size) })
 {
 	setTexture(texture);
