+++ SquareParticle.h
@@ -4,5 +4,5 @@
 class SquareParticle : public Particle
 {
 public:
-	SquareParticle(float speed, sf::Vector2f direction, float size, sf::Texture& texture);
+	SquareParticle(int speed, sf::Vector2f direction, int size, sf::Texture& texture);
 };
